added success/error messages throughout

+++ friend-of-bill/client/src/components/NightlyReview.js
@@ -2,8 +2,12 @@ import React from "react";
 import { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
 import { selectUser } from "../features/userSlice";
+import { useDispatch } from "react-redux";
+import { selectMessage, setMsg } from '../features/messageSlice';
 import { Link } from 'react-router-dom';
 import { useHistory } from "react-router-dom";
+
+import AlertMessage from "./AlertMessage";
 import "./NightlyReview.css";
 
 const NightlyReview = () => {
@@ -11,8 +15,9 @@ const NightlyReview = () => {
 
   //Bring in user info from store
   const user = useSelector(selectUser);
+  const dispatch = useDispatch();
+  const message = useSelector(selectMessage);
   const history = useHistory();
-  console.log(user.id);
 
   //if no user is logged in, redirect them back to login page
   if (user.id === undefined) {
@@ -29,7 +34,8 @@ const NightlyReview = () => {
       })
       .then((data) => {
         console.log(data.review);
-        setReviews(data.review.reverse());  //I want the most recent review to be first for the user. Easiest way to do that for me way to just reverse the array being sent from my backend
+        //I want the most recent review to be first for the user. Easiest way to do that for me way to just reverse the array being sent from my backend
+        setReviews(data.review.reverse());
       })
       .catch((err) => {
         //this is where I will display message to user
@@ -38,6 +44,8 @@ const NightlyReview = () => {
   }, []);
 
  const handleDelete = id => {
+  const confirm = window.confirm('Are you sure you want to delete this item?');
+  if(confirm) {
    fetch(`/api/reviews/${id}`, {
      method: "DELETE",
      headers: {
@@ -46,19 +54,21 @@ const NightlyReview = () => {
      body: JSON.stringify(user)
    }).then(res => {
      if(!res.ok) {
-       throw new Error(`Error getting delete endpoint`)
+       throw new Error(`Something went wrong when trying to delete this item. Please try again later!`)
      }
      return res.json();
    }).then(data => {
     setReviews(data.reviews.reverse());
+    dispatch(setMsg({ msg: 'Nightly Review sucessfully deleted!', err: false }));
    }).catch(err => {
-     console.log(err);
+    dispatch(setMsg({ msg: err.message, err: true }))
    })
+  }
  }
 
   return (
     <div className="nightlyReview">
-      <h1>Nightly Reviews</h1>
+      { message && <AlertMessage marginTop/> }
       <Link className="nightly-review-link custom-btn btn--green" to="/reviews/new">New Review</Link>
       <div className="nightlyReview__container">
       { reviews.map(el => (
